requirements.txt
PyJWT==2.8.0
cryptography==41.0.7
rich==13.7.0
pyperclip==1.8.2